Refactor, renaming and such...
- moved some stuff into magix::compile and magix::execute
- some more comments to explain
- minor refactor here and there, mainly moving methods into appropiate
  header
- fixed some warnings

+++ gdextension/cpp/magix_vm/src/magix_vm/doctest_helper.hpp
@@ -87,7 +87,7 @@ doctest_range_eq_impl(
 auto
 doctest_bytestring_eq_impl(
     const char *file,
-    size_t line,
+    int line,
     const doctest::String &name_a,
     const doctest::String &name_b,
     magix::span<const std::byte> bytes_a,
